Refactored AccountsRowHeight feature to webpack framework.
