--- source/common/res/features/accounts-row-height/settings.json
@@ -1,18 +1,0 @@
-{
-         "name": "accountsRowHeight",
-         "type": "select",
-      "default": "0",
-      "section": "accounts",
-        "title": "Height of Rows in Account Register",
-  "description": "Change the height of transaction rows so more of them are displayed on the screen.",
-      "options": [
-                    { "name": "Default", "value": "0" },
-                    { "name": "Compact", "value": "1" },
-                    { "name": "Slim", "value": "2" }
-                 ],
-      "actions": {
-                    "1": [ "injectCSS", "compact.css", "injectScript", "main.js" ],
-                    "2": [ "injectCSS", "slim.css", "injectScript", "main.js" ]
-                 }
-}
-
